requirements.txt
torch>=2.1.0
torchvision>=0.16.2
transformers>=4.36.0
accelerate>=0.26.0
datasets>=2.15.0
google-cloud-aiplatform
huggingface-hub
bitsandbytes
peft
